requirements.txt
fastapi==0.116.1
httpx==0.28.1
openai==1.99.9
pydantic==2.11.7
python-dotenv==1.1.1
uvicorn==0.35.0
